refactor(diagram): tighten the house-root module surface
Review findings on the consolidation.

HOUSE_STYLE and naturalSize were exported but only houseRoot in the same
file reads them, so both lose the export; nothing outside the module
imported either, tests included.

A prior style already ending in `;` produced `;;` when the cap was merged
after it. The separator is normalised before appending, with a test case.

drawio-render.ts took SvgError from d2-svg.ts, which was the last
wrong-direction import left: it now takes it from house-root.ts, where
the class lives. The d2-svg.ts re-export stays for d2-render.ts and the
two test files that still use it.

Also drops a trailing blank line at the end of house-root.ts.

Refs #456
